+++ internal/cmd/clone.go
@@ -8,13 +8,13 @@ import (
 	"repo/internal/hoster/gitlab"
 	"repo/internal/say"
 	"sort"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/spf13/cobra"
 )
 
+var cloneStyle string
 var cloneTopics []string
 var cloneExcludePatterns []string
 var cloneIncludePatterns []string
@@ -24,6 +24,7 @@ var cloneStarred bool
 
 func init() {
 	rootCmd.AddCommand(cloneCmd)
+	cloneCmd.Flags().StringVarP(&cloneStyle, "style", "y", "flat", "Either repositories are cloned 'flat' into the root-dir, or 'recursive' using the groups as directories.")
 	cloneCmd.Flags().StringSliceVarP(&cloneTopics, "topic", "t", nil, "Topics (aka tags/labels) to be filtered. Multiple topics are possible (and).")
 	cloneCmd.Flags().StringSliceVarP(&cloneExcludePatterns, "exclude", "e", nil, "Regex-pattern not to be matched for the path. Multiple patterns are possible (and).")
 	cloneCmd.Flags().StringSliceVarP(&cloneIncludePatterns, "include", "i", nil, "Regex-pattern that needs to be matched for the path. Multiple patterns are possible (and).")
@@ -38,7 +39,7 @@ var cloneCmd = &cobra.Command{
 	Args:  validateConditions(cobra.ExactArgs(1), validateArgGitDir(0, false, true)),
 	Run: func(cmd *cobra.Command, args []string) {
 		defer say.Timer(time.Now())
-		dirReposRoot := args[0]
+		dirReposRoot := getAbsoluteRepoRoot(args[0])
 
 		hoster, err := gitlab.MakeHoster()
 		handleFatalError(err)
@@ -51,39 +52,37 @@ var cloneCmd = &cobra.Command{
 			IncludePatterns: cloneIncludePatterns,
 		})
 		sort.Slice(repos, func(i, j int) bool {
-			return repos[i].Name < repos[j].Name
+			return repos[i].PathWithNamespace < repos[j].PathWithNamespace
 		})
 
 		repos = filterExisting(dirReposRoot, repos)
 		cloneAll(dirReposRoot, repos)
 	},
+	PreRunE: validateFlags,
 }
 
 func filterExisting(dirReposRoot string, repos []h.HosterRepository) (result []h.HosterRepository) {
 	for _, r := range repos {
-		dirName := determineDirectoryName(r.SshUrl)
-		dirRepository := path.Join(dirReposRoot, dirName)
+
+		var dirRepository string
+		if Style == styleFlat {
+			dirRepository = path.Join(dirReposRoot, r.Path)
+		} else if Style == styleRecursive {
+			dirRepository = path.Join(dirReposRoot, r.PathWithNamespace)
+		}
 
 		_, err := os.Stat(dirRepository)
 		if os.IsNotExist(err) {
 			result = append(result, r)
 		} else {
-			say.Verbose("Repository exists already: %s ", dirName)
+			say.Verbose("Repository exists already: %s ", r.PathWithNamespace)
 		}
 	}
 	return
 }
 
-func determineDirectoryName(sshUrl string) string {
-	last := sshUrl[strings.LastIndex(sshUrl, "/")+1:]
-	if strings.HasSuffix(last, ".git") {
-		last = last[:len(last)-4]
-	}
-	return last
-}
-
 func cloneAll(dirReposRoot string, repos []h.HosterRepository) {
-	tasks := make(chan string)
+	tasks := make(chan h.HosterRepository)
 	var wg sync.WaitGroup
 	counter := int32(0)
 	for i := 0; i < getParallelism(cloneParallelism); i++ {
@@ -92,22 +91,29 @@ func cloneAll(dirReposRoot string, repos []h.HosterRepository) {
 	}
 
 	for _, repo := range repos {
-		tasks <- repo.SshUrl
+		tasks <- repo
 	}
 
 	close(tasks)
 	wg.Wait()
 }
 
-func clone(dirReposRoot string, counter *int32, total int, tasks chan string, wg *sync.WaitGroup) {
+func clone(dirReposRoot string, counter *int32, total int, tasks chan h.HosterRepository, wg *sync.WaitGroup) {
 	defer wg.Done()
-	for sshUrl := range tasks {
-		dirName := determineDirectoryName(sshUrl)
-		err := gitclient.Clone(dirReposRoot, dirName, sshUrl)
+	for repo := range tasks {
+
+		var dirTarget string
+		if Style == styleFlat {
+			dirTarget = repo.Path
+		} else if Style == styleRecursive {
+			dirTarget = repo.PathWithNamespace
+		}
+
+		err := gitclient.Clone(dirReposRoot, dirTarget, repo.SshUrl)
 		if err != nil {
-			say.ProgressError(counter, total, err, dirName, "- Unable to clone")
+			say.ProgressError(counter, total, err, repo.PathWithNamespace, "- Unable to clone")
 		} else {
-			say.ProgressSuccess(counter, total, dirName, "")
+			say.ProgressSuccess(counter, total, dirTarget, "")
 		}
 	}
 }