+++ internal/hoster/gitlab/gitlab.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"os"
 	"regexp"
+	"slices"
 	"strconv"
 	"time"
 
@@ -15,25 +16,30 @@ import (
 	"repo/internal/say"
 	"repo/internal/util"
 
+	"github.com/xanzy/go-gitlab"
 	gg "github.com/xanzy/go-gitlab"
 )
 
-const REPOW_GITLAB_API_TOKEN = "REPOW_GITLAB_API_TOKEN"
-const GITLAB_API_TOKEN = "GITLAB_API_TOKEN"
+const (
+	REPOW_GITLAB_API_TOKEN = "REPOW_GITLAB_API_TOKEN"
+	GITLAB_API_TOKEN       = "GITLAB_API_TOKEN"
 
-const REPOW_GITLAB_DOWNLOAD_RETRIES = "REPOW_GITLAB_DOWNLOAD_RETRIES"
-const GITLAB_DOWNLOAD_RETRIES = "GITLAB_DOWNLOAD_RETRIES"
-const GITLAB_DOWNLOAD_RETRIES_DEFAULT = 6 // lower values didn't solve the issue
+	REPOW_GITLAB_DOWNLOAD_RETRIES   = "REPOW_GITLAB_DOWNLOAD_RETRIES"
+	GITLAB_DOWNLOAD_RETRIES         = "GITLAB_DOWNLOAD_RETRIES"
+	GITLAB_DOWNLOAD_RETRIES_DEFAULT = 6 // lower values didn't solve the issue
+
+	REPOW_GITLAB_HOST = "REPOW_GITLAB_HOST"
+)
 
 func MakeHoster() (*Gitlab, error) {
 	result := &Gitlab{}
-	value := util.GetEnv(REPOW_GITLAB_API_TOKEN, util.GetEnv(GITLAB_API_TOKEN, ""))
-	if value == "" {
-		return result, errors.New("The " + REPOW_GITLAB_API_TOKEN + " or " + GITLAB_API_TOKEN + " environment-variable has to be set")
+	valueApiToken := util.GetEnv(REPOW_GITLAB_API_TOKEN, util.GetEnv(GITLAB_API_TOKEN, ""))
+	if valueApiToken == "" {
+		return result, errors.New("the " + REPOW_GITLAB_API_TOKEN + " or " + GITLAB_API_TOKEN + " environment-variable has to be set")
 	}
 
 	var errClient error
-	result.client, errClient = gg.NewClient(value)
+	result.client, errClient = gg.NewClient(valueApiToken, gitlab.WithBaseURL("https://"+result.Host()))
 	if errClient != nil {
 		return nil, errClient
 	}
@@ -45,7 +51,7 @@ type Gitlab struct {
 }
 
 func (g Gitlab) Host() string {
-	return "gitlab.com"
+	return util.GetEnv(REPOW_GITLAB_HOST, "gitlab.com")
 }
 
 func (g Gitlab) Repositories(options hoster.RequestOptions) []hoster.HosterRepository {
@@ -77,9 +83,14 @@ func (g Gitlab) Repositories(options hoster.RequestOptions) []hoster.HosterRepos
 			total++
 
 			if matches(options, project.PathWithNamespace, project.TagList) {
+				//_, pathWithoutNamespace, _ := strings.Cut(project.PathWithNamespace, "/")
+				//say.Info("\npath: %s (was: %s)", rootlessPath, project.PathWithNamespace)
 				repos = append(repos, hoster.HosterRepository{
-					Id:     project.ID,
-					Name:   project.Name,
+					Id:                project.ID,
+					Name:              project.Name,
+					Path:              project.Path,
+					PathWithNamespace: project.PathWithNamespace,
+					//PathWithoutNamespace: pathWithoutNamespace,
 					Topics: project.TagList,
 					SshUrl: project.SSHURLToRepo})
 			}
@@ -195,7 +206,7 @@ func (g Gitlab) Validate(repo model.RepoMeta, optionalContacts bool) []error {
 	// gitlab features
 	gl := repo.RepoYaml.Gitlab
 	allowed := []string{"", "enabled", "private", "disabled"}
-	if gl.WikiAccessLevel != nil && !util.IsInSlice(*gl.WikiAccessLevel, allowed...) {
+	if gl.WikiAccessLevel != nil && !slices.Contains(allowed, *gl.WikiAccessLevel) {
 		errs = append(errs, errors.New(fmt.Sprintf("WikiAccessLevel must be one of: %s", allowed)))
 	}
 	//TODO check more gitlab values