+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.3.0] - 2025-05-18
+
+### Added
+* The `clone` command does support recursive structures for groups using the `style` argument. An environment-variable `REPOW_STYLE` can also be set, to set the default behaviour.
+* Providing an experimental windows binary.
+* Support for self-hosted Gitlab installations. The Gitlab-Host can be changed via the environment Variable `REPOW_GITLAB_HOST`.
+
+
 ## [0.2.3] - 2022-11-09
 
 ### Added
@@ -83,7 +91,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/galan/repow/compare/v0.2.3...HEAD
+[Unreleased]: https://github.com/galan/repow/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/galan/repow/compare/v0.2.3...v0.3.0
 [0.2.3]: https://github.com/galan/repow/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/galan/repow/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/galan/repow/compare/v0.2.0...v0.2.1